@@ -9,6 +9,9 @@
 namespace StatusLib;
 
 use DomainException;
+use Psr\Container\ContainerInterface;
+
+use function sprintf;
 
 /**
  * Service factory for the StatusLib TableGateway
@@ -23,6 +26,10 @@ use DomainException;
  */
 class TableGatewayFactory
 {
+    /**
+     * @param ContainerInterface $services
+     * @return TableGateway
+     */
     public function __invoke($services)
     {
         $db    = 'Db\StatusLib';
@@ -32,8 +39,8 @@ class TableGatewayFactory
             switch (isset($config['statuslib'])) {
                 case true:
                     $config = $config['statuslib'];
-                    $db     = isset($config['db']) ? $config['db'] : $db;
-                    $table  = isset($config['table']) ? $config['table'] : $table;
+                    $db     = $config['db'] ?? $db;
+                    $table  = $config['table'] ?? $table;
                     break;
                 case false:
                 default:
