@@ -8,19 +8,22 @@
 
 namespace StatusLib;
 
+use stdClass;
+use Traversable;
+
 /**
  * Interface for StatusLib mappers
  */
 interface MapperInterface
 {
     /**
-     * @param array|\Traversable|\stdClass $data 
+     * @param array|Traversable|stdClass $data
      * @return Entity
      */
     public function create($data);
 
     /**
-     * @param string $id 
+     * @param string $id
      * @return Entity
      */
     public function fetch($id);
@@ -31,14 +34,14 @@ interface MapperInterface
     public function fetchAll();
 
     /**
-     * @param string $id 
-     * @param array|\Traversable|\stdClass $data 
+     * @param string $id
+     * @param array|Traversable|stdClass $data
      * @return Entity
      */
     public function update($id, $data);
 
     /**
-     * @param string $id 
+     * @param string $id
      * @return bool
      */
     public function delete($id);
